Patient list pagination: Total results cached for proper navigation in pagination
